k8s: e2e: remove dash from cri-o in skipped_tests_e2e.yaml
We use `CRI_RUNTIME` to specify which runtime interface we are
using to launch k8s and to know which tests we need to skip for e2e
testing.
Specifically for cri-o, we use `CRI_RUNTIME=crio`, so we
need to remove the dash (-) from cri-o in the skipped_tests_e2e.yaml
file, so that the tests listed there are properly skipped.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>