ci: lib.sh: set PATH for sudo even in no-go case
If the scripts failed to find go or a GOPATH they would follow a different
path and invoke `sudo -E`, but not with the PATH=$PATH argument. This
normally works on oru CIs, but on Zuul, where the old distro go version
will be in the sudo path first, it fails with a 'go too old' error.
Pass the CI env/user PATH into the `sudo` to ensure we pick up
the correct golang.

Fixes: #1479

Signed-off-by: Graham Whaley <graham.whaley@intel.com>