kata-manager: Check for dependencies earlier
The `kata-manager` script needs to download `git` repos. But if `git`
isn't available, it can fall back to using `curl` or `wget`. However, if
none of these tools are available, it should exit before any changes are
made to the system.

Note: The script _could_ install one of the tools of course. However,
all known Linux distro "default installs" provide at least one of these
tools, so this scenario should only be hit on very minimal systems. In
that case, install one of the tools the error mentions and re-run
`kata-manager`.

Fixes #1502.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>