CI: Add markdown checker
Add a simple tool that checks markdown files to ensure all links are
valid.

Updated the static checker script to call this tool to ensure that
documents with broken links are not merged.

Note that when the tool detects a broken link, it is often able to
suggest the correct link name so looking at its output is beneficial.

Fixes #1485.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>