integration/docker: restart container if hotplug cpu test fails
Sometime the hypervisor/kernel is not able to allocate the vcpus,
the test waits for around 25 seconds, but the vcpus is never added.
To avoid random failures in the CI, the test run the container, start the
test that checks the actual number of vcpus, if it fails then the container
is removed and the loop starts again, until the test success or the maximum
number of retries have been reached.

fixes #1529

Signed-off-by: Julio Montes <julio.montes@intel.com>