pkglib: change short commit length
change short commit length to be the same to OBS.

Now that runtime has a strict depedency version from packages.

Like qemu-lite:

Now:
qemu-lite = 2.11.0+git.87517af

Before:
qemu-lite >= 2.11.0+git.87517af

The rpm fails because the real package version of qemu lite is

2.11.0+git.87517afd72

The commit length comes from the format of OBS '%h'

This change the shortcommit length to be the same to OBS
and runtime dependencies and packages that include git commit
as version use the same format.

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>