check-markdown: Add list sub-commands
Add the ability to list headings and links with new sub-commands
(`list headings` and `list links` respectively).

The default output format is almost a raw golang dump (however it is
human-readable). Alternatively, you can specify `--format tsv` for
tab-separated fields (or use `--separator` if you wish to use an
alternative separator character such as `:` for CSV).

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>