CI: Add unreferenced doc check
Add a new test to the CI static check script to look for markdown
documents that are not referenced by any other. With a small number of
exceptions, most markdown files should be linked to from other markdown
documents in the same repository to allow all in-repository
documentation to be discovered by link navigation.

Fixes: #1588.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>