labelsdb: Add review label
Added a new `needs-review` label. The idea is that the new GitHub issue
templates can all add this label when issues are created. The Triage
Team will then review the new issues, add any missing labels and
*remove* the `needs-review` label.

The `needs-review` label thus provides a simple way to identify issues
that have not been looked at yet.

The label could also be applied to PRs in some circumstances.

Note that the original name of a `needs-triage` label was rejected
since:

- Triage is an unusual word which may be confusing to users.
- Triage implies "initial" so would only really make sense for issues.

Fixes: #1658.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>