CI: Add arch options to static check script
The static check script runs a number of checks, but some (one) are
architecture-specific. Add in three new options to allow the behaviour
of the script to be modified:

- `--list`: List all available check functions.
- `--no-arch`: List or run only architecture-agnostic check functions.
- `--only-arch`: List or run only architecture-specific check functions.

Note that now the script searches itself for the list of functions to
run rather than hard-coding the list of functions to run. This
necessitated changing the name of each check function to
`static_check_*()` to ensure only functions in this file are considered.

Fixes: #1673.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>