integration/k8s: expose ip: update image, fix test
The expose IP test was using the node-hello image, which is now
marked as deprecated, and was also very large. Replace that
image with the hello-app image, which is much smaller.

The test was also checking it could curl from the pod
IP addresses, whereas we were deploying a loadbalancer service,
and really want to be testing we can curl from that IP instead.
Update the test to do that.
Note, it does not test on the 'external IP', as that may not
get allocated, depending on the system setup.

Fixes: #1692

Signed-off-by: Graham Whaley <graham.whaley@intel.com>