integration/k8s: block test: use alpine rather than debian
The block volume test was using a Debian image to get access
to mkfs.ext4. We can get that using alpine if we apk install
an extra package, which should give us an overall smaller
image footprint etc.

The original test invoked `mkfs` with an ext4 option, but
the generic `mkfs` is not available in the Alpine package,
so invoke `mkfs.ext4` directly instead.

Fixes: #1695

Signed-off-by: Graham Whaley <graham.whaley@intel.com>