cmd: Add spell checker tool
Added a new spell checker tool based on Hunspell. The tool will look for
spelling mistakes in a document using standard English dictionaries and
a custom dictionary to ensure well-known technical terms are not
rejected.

A number of simplifications are performed on the document before it is
checked to minimise false-positives and maximise reliability/usability.
See the new README for further details.

Note that the tool is not supported on OSX due to the difficulty of
installing hunspell dictionaries.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>