integration/docker: move user_test to busybox image
The user_test was the only test using the postgres image.
That image is very large, and there is no particular reason
the test needs that specific image, it just looks like a legacy
historical thing.
Move the test to use busybox, and adapt the tests to cater for
the minor changes in the group file and `id` output.

Fixes: #1700

Signed-off-by: Graham Whaley <graham.whaley@intel.com>