ci: jenkins: metrics: add metrics call direct to jenkins script
Currently we call the jenkins script to set up the system, and then call
the run_metrics_PR_ci.sh script from the metrics jenkins job files to
run the actul metrics.
This flow does not 'understand' the new ci fast path, and then fails.
The best fix is to move the call to the metrics script into the main
jenkins file, and remove it from the jenkins job files.

Fixes: #1835

Signed-off-by: Graham Whaley <graham.whaley@intel.com>