vendor: Unify Gopkg files and vendor folder within the repository
This repository contains multiple Gopkg.* files which defines
dependencies for each stand-alone application within this repository.
This was posing an issue where there was no predictable way of
confirming that vendor folder is in sync with the versions defined in
Gopkg files. dep tool does not handle multiple sub folder definitions
and is not a feature that is in their pipeline. This effort is to unify
all the dependencies to allow `dep check` on the repository. Check
https://github.com/golang/dep/issues/985 for dep issues with nesting.

Fixes: #1794
Fixes: #1733
Signed-off-by: Ganesh Maharaj Mahalingam <ganesh.mahalingam@intel.com>