dictionary: Add missing distro terms
Add "MacOS" and "OpenSUSE" to the distros fragment for the custom
dictionary.

Both these terms already existed in distros fragments file, but with
different capitalisation, correct for each term.

Although `hunspell(1)` will automatically accept the capitalised
versions of custom words specified in the dictionary in all lower-case,
it will *not* accept capitalised custom words starting with a lower-case
letter but containing an upper-case one, like both of the previously
specified terms.

Fixes: #1866.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>