modules: Verification for go modules
While we move to go modules, perform the dep check for repos
that still use dep.
Run `go mod verify` instead for go modules.
Note, this just verifies the integrity of modules in the local
cache. We would have instead wanted to verify the vendored code
here, but that is still not supported.
https://github.com/golang/go/issues/27348

Fixes #1879

Signed-off-by: Archana Shinde <archana.m.shinde@intel.com>