cleanup_k8s: ensure execution of cleanup with `|| true`
there are several scenarios where cleanup_env.sh wouldn't be invoked
or be interrupted in the middle.
Sometimes, we failed on `kubeadm init`. on such point, the cluster
network may not be initialized, which means no `cni0` interface, regardless
of `flannel` interface. Hence, cleanup will exit in the middle.
So adding `|| true` ensures that the resulting compound command
always exits with status zero.

Fixes: #1913

Signed-off-by: Penny Zheng <penny.zheng@arm.com>