virtiofs: Skip package manager test for dnf and yum
dnf and yum update tests are failing when running with
virtiofs. More info can be found on #2008.

While this is being fixed, lets skip it to have a green CI.

In addition, unskip the yum test running with 9p, which was
not working in the past but now behaves correctly.

Fixes: #2013.
Fixes: #1270.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>