release: Kata Containers 1.9.4
The branch 1.9 was created after Kata VERSION is created, for the
current release process this is repository is not needed but CI will
fail doing checkout for this repository, also CI for releases will check
that all repos are in the same version so it is only needed for CI
correctness.

- version: Add VERSION file
- ci: Add minimal makefile to use central go test script
- netlink: pull out netlink as library crate.
- Fixup workflow 103

8d60612 version: Add VERSION file
a5192a1 netlink: pull out netlink as library crate.
3881c06 ci: Add minimal makefile to use central go test script
1c57665 workflows: make sure we build the experimental kernel, CLH
cbd5fa0 workflows: fix step output usage

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>