ci: static check: spellcheck: check all docs
Rather than fail the spell check on the first failure, continue to
check all the changed documents, and report the global fail at the
end. This may help reduce CI/review cycle time if/when multiple docs
are submitted in a single PR.

Also spit out a URL link to the documentation repo docs in the case
of failure, to help direct authors to guidance docs.

Fixes: #2227

Signed-off-by: Graham Whaley <graham.whaley@intel.com>