k8s: Lower the number of CPUs on CPU test
We need to lower the request of CPUs on the pod-number-cpu.yaml
which is used to run our CPU test. This is required as we now
have enabled PodOverhead and our CI VMs are small.

In addition, increase the sleep time that is run inside one of the
containers, so that we are sure that we can measure the number
of CPUs that it is using.

Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>