pkgs: spec do not check for 'not in final'
Do not check for 'not in final' in spec creation, the logic
to fully validate is longer that just one grep.

Next should:

Use the same script build-kernel.sh to generate spec and validate it.

For now is still safe as CI will run all the build-kernels.sh to verify
the resulting config.

Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>