ci: static: markdown: make list var func local
The `list` var in the skip_paths function was not local, thus
afiact, rendering it global. Thus, when skip_paths was called
more than once, the list grew and grew, and included the unfiltered
files, including vendor files.

Fix by making it local, so it is built afresh on each function
invocation.

Fixes: #2338

Signed-off-by: Graham Whaley <graham.whaley@intel.com>