Makefile: Run docker tests sequentially in CentOS 7
Docker tests fail randomly in Centos 7 if they run in parallel.
The systemd service in the guest uses `fstat` to build a map of unit files.
For a strange reason, the `fstat` operation on pmem devices takes too much
time, looks like this could be a bug in the host kernel, because it is not
reproducible in CentOS 8.
Run docker test sequentially to avoid random failures.

fixes #2355

Signed-off-by: Julio Montes <julio.montes@intel.com>