ci: Bring back docker integration tests since yq v3.1
All docker integration tests have not been running since yq was updated
to v3.1 (for at least 12 days). The root cause is that yq v3.1 is no
longer returning 'null' as empty strings. To be safe, we keep checks on
both 'null' and '""' (empty string). Also, we observed some regression
test failures with this patch (at least for CIs on ubuntu-initrd, clh,
and Fedora31).

Fixes: #2368

Signed-off-by: Bo Chen <chen.bo@intel.com>
Signed-off-by: Jose Carlos Venegas Munoz <jose.carlos.venegas.munoz@intel.com>
Signed-off-by: Salvador Fuentes <salvador.fuentes@intel.com>